Load characters method

+++ AverageRaiderIoScore/Character.cs
@@ -5,7 +5,7 @@
         public string Name { get; set; }
         public string Realm { get; set; }
         public Region Region { get; set; } = Region.eu;
-        public int RaiderIoScore { get; set; }
-        public int ItemLvl { get; set; }
+        public double RaiderIoScore { get; set; }
+        public double ItemLvl { get; set; }
     }
 }
